Move Script Tag to Head & Use Defer
Move Script Tag to Head & Use Defer as that is the preferred way of
loading JavaScript files since it can fetch and execute the JavaScript
as the HTML file is loading. This results in faster performance since
the HTML and JavaScript are loading simultaneously, and the JavaScript
can load in the order the developer wants it to load.