Add Partially Update DOM Function For Servings
Add a function to partially update the DOM by using the built-in visual
DOM, as the page would flicker each time it was updated. And the DOM
updated the page too much, which would slow down the render; this way,
the parts that have changed are the only ones that get updated. However,
this algorithm should not be used in big projects as it has a bad Big O.